requirements.txt
django-recaptcha==4.0.0
django-widget-tweaks==1.5.0
asgiref==3.8.1
Django==5.1.4
django-cors-headers==3.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
typing_extensions==4.12.2